컬럼명 수정

+++ backend/jobplatform/src/main/java/jobplatform/fo/user/domain/mapper/M_MypageMapper.java
@@ -35,10 +35,10 @@ public interface M_MypageMapper {
 	
 	
 //	포지션 제안 수락 여부 Check
-	public String getPstnPrpslAcceptYN(int mbr_sq);
+	public String getPstnPrpslAcceptYN(Long mbr_sq);
 
 // 	포지션 제안 수락 여부 변경
-	public int updatePstnPrpslAcceptYN(@Param("mbr_sq") int mbr_sq, @Param("pstn_prpsl_accept_yn") String pstn_prpsl_accept_yn);
+	public int updatePstnPrpslAcceptYN(@Param("mbr_sq") Long mbr_sq, @Param("pstn_prpsl_accept_yn") String pstn_prpsl_accept_yn);
 
 //	지역 정보 가져오기
 	public List<Map<String, Object>> getAreaCodes();
@@ -47,22 +47,22 @@ public interface M_MypageMapper {
 	public List<Map<String, Object>> getJobCodes();
 
 // 	설정했던 지역 검색
-	public List<Integer> getSelectedAreas(int mbr_sq);
+	public List<Integer> getSelectedAreas(Long mbr_sq);
 
 // 	선택했던 지역 초기화
-	public int resetSelectedAreas(int mbr_sq);
+	public int resetSelectedAreas(Long mbr_sq);
 
 // 	선택한 지역 INSERT
-	public int insertSelectAreas(@Param("mbr_sq") int mbr_sq, @Param("areaList") List<Integer> areaList);
+	public int insertSelectAreas(@Param("mbr_sq") Long mbr_sq, @Param("areaList") List<Integer> areaList);
 
 // 	설정했던 직업 검색
-	public List<Integer> getSelectedJobs(int mbr_sq);
+	public List<Integer> getSelectedJobs(Long mbr_sq);
 
 // 	선택했던 직업 초기화
-	public int resetSelectedJobs(int mbr_sq);
+	public int resetSelectedJobs(Long mbr_sq);
 
 // 	선택한 직업 INSERT
-	public int insertSelectJobs(@Param("mbr_sq") int mbr_sq, @Param("jobList") List<Integer> jobList);
+	public int insertSelectJobs(@Param("mbr_sq") Long mbr_sq, @Param("jobList") List<Integer> jobList);
 
 //	포지션 제안 받은 공고 리스트 출력(해당 페이지 데이터)
 	public List<M_JobPosting_pp> getPPJopPostingData(@Param("rsm_sq") int rsm_sq, @Param("limit") int limit, @Param("offset") int offset);
@@ -86,7 +86,7 @@ public interface M_MypageMapper {
 	
 // 	@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@tools@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
 //	대표이력서 rsm_sq 구하는 메서드
-	public int getRRsm_sq(int mbr_sq);
+	public int getRRsm_sq(Long mbr_sq);
 //	지역 코드 -> 지역 이름
 	public List<String> areaCodeToName(@Param("areas") List<Integer> areas);
 //	직업 코드 -> 직업 이름
